refactor: move limits tests to enterprise
